requirements.txt
aiohttp==3.13.2
qrcode==8.2
pillow==12.0.0
pynacl==1.6.1
git+https://github.com/ezzygarmyz/hdwallet.git
certifi==2025.11.12